build: ondemand simulatr with tests
