+++ .github/workflows/publish_release.yml
@@ -20,8 +20,8 @@ jobs:
           # version.
           fetch-depth: 0
 
-      - name: Run publish release action
-        uses: ./subtrees/z_quantum_actions/actions/publish-release
-
       - name: run cloud tests
         uses: ./cloudcoverage
+
+      - name: Run publish release action
+        uses: ./subtrees/z_quantum_actions/actions/publish-release
