+++ src/orquestra/integrations/braket/simulator/OnDemandSimulator.py
@@ -1,23 +1,16 @@
 ################################################################################
 # © Copyright 2021-2022 Zapata Computing Inc.
 ################################################################################
-from typing import Any, List, Sequence
 
-import numpy as np
 from boto3 import Session  # type: ignore
 from braket.aws import AwsDevice
 from braket.aws.aws_session import AwsSession
+from orquestra.quantum.api.backend import QuantumSimulator
 from orquestra.quantum.circuits import Circuit
 from orquestra.quantum.distributions import MeasurementOutcomeDistribution
 from orquestra.quantum.measurements import Measurements
 
-from orquestra.integrations.braket.conversions import export_to_braket
-
-from ._base import (
-    BraketBasedSimulator,
-    _get_arn,
-    _get_measurement_from_braket_result_object,
-)
+from ._base import _get_arn, _run_circuit_and_measure
 
 
 class BraketOnDemandSimulator:
@@ -84,59 +77,20 @@ class BraketOnDemandSimulator:
         self.number_of_circuits_run += 1
         self.number_of_jobs_run += 1
 
-        braket_circuit = export_to_braket(circuit)
+        return _run_circuit_and_measure(self, circuit, n_samples)
 
-        result_object = self.simulator.run(braket_circuit, shots=n_samples).result()
-
-        return _get_measurement_from_braket_result_object(result_object)
-
-    def run_circuitset_and_measure(
-        self, circuits: Sequence[Circuit], n_samples: Sequence[int]
-    ) -> List[Measurements]:
-        """Run a set of circuits and measure a certain number of bitstrings.
-
-        It may be useful to override this method for backends that support
-        batching.
-
-        Args:
-            circuits: The circuits to execute.
-            n_samples: The number of samples to collect for each circuit.
-        """
-        measurement_set: List[Measurements]
-
-        if not self.supports_batching:
-            measurement_set = []
-            for circuit, n_samples_for_circuit in zip(circuits, n_samples):
-                measurement_set.append(
-                    self.run_circuit_and_measure(
-                        circuit, n_samples=n_samples_for_circuit
-                    )
-                )
-
-            return measurement_set
-        else:
-            self.number_of_circuits_run += len(circuits)
-            if isinstance(self.batch_size, int):
-                self.number_of_jobs_run += int(np.ceil(len(circuits) / self.batch_size))
-
-            # This value is only returned so that mypy doesn't complain.
-            # You can remove this workaround when we reimplement counter increments in
-            # a more type-elegant way.
-            measurement_set = []
-            return measurement_set
+    def run_circuitset_and_measure(self, circuits, n_samples):
+        return QuantumSimulator.run_circuitset_and_measure(self, circuits, n_samples)
 
     def get_measurement_outcome_distribution(
         self, circuit: Circuit, n_samples: int
     ) -> MeasurementOutcomeDistribution:
         """Calculates a measurement outcome distribution.
-
         Args:
             circuit: quantum circuit to be executed.
             n_samples: number of samples to collect
-
         Returns:
             Probability distribution of getting specific bistrings.
-
         """
         # Get the expectation values
         measurements = self.run_circuit_and_measure(circuit, n_samples)
